+++ examples/run_multiagent_coding.py
@@ -1,4 +1,4 @@
-from experiments.multiagent_coding.smolagents.multiagent_coding_smolagents import MultiAgentCoding
+from multiagent_coding.smolagents.multiagent_coding_smolagents import MultiAgentCoding
 
 coding = MultiAgentCoding()
 result = coding.run("Write a Python function to calculate factorial recursively.")
