+++ examples/run_synthesis.py
@@ -1,4 +1,4 @@
 import asyncio
-from experiments.synthesis.test_synthesis import test
+from synthesis.test_synthesis import test
 
 asyncio.run(test())
