+++ multiagent_coding/noncoding_examples_for_reference/noncoding_multiagent_example_smolagents.py
