+++ synthesis/synthesis.py
@@ -114,3 +114,5 @@ class SynthesisModel:
         """
         loop = asyncio.get_event_loop()
         return await loop.run_in_executor(None, model_func, prompt)
+
+
