+++ terminal_multiagent_coding.py
@@ -1,5 +1,5 @@
 import sys
-from experiments.multiagent_coding.smolagents.multiagent_coding_smolagents import MultiAgentCoding
+from multiagent_coding.smolagents.multiagent_coding_smolagents import MultiAgentCoding
 
 def main():
     """
